+++ NEWS.md
@@ -1,3 +1,5 @@
+- New `with_output_sink()` and `with_message_sink()` (#24).
+
 # 1.0.0
 
 First Public Release